fixed npm start bug
fixed registerSlashCommand freezing because child_process is being opened
