+++ src/registerSlashCommand.ts
@@ -3,6 +3,7 @@ const debug = require("debug")("Bot:registerSlashCommand");
 import { REST, RESTPostAPIChatInputApplicationCommandsJSONBody, Routes } from 'discord.js';
 import { CLIENT_ID, BOT_GUILD_IDS, DISCORD_TOKEN } from '@config';
 import { allCommands } from "@shared/commands";
+import { yoloService } from '@shared/YoloService';
 
 if(DISCORD_TOKEN === undefined){
     console.error("Error: Discord token is not defined!");
@@ -34,6 +35,8 @@ const rest = new REST({ version: '10' }).setToken(DISCORD_TOKEN);
             debug(`List of all commands: ${slashCommands.map(cmd => cmd.name).join(" | ")}`);
         }
         console.log("All commands were registered sucessfuly!");
+
+        yoloService.shutdown();
 	} catch (error) {
 		// And of course, make sure you catch and log any errors!
 		console.error(error);
