+++ src/shared/YoloService.ts
@@ -1,0 +1,3 @@
+import { YOLOService } from "@library";
+
+export const yoloService = new YOLOService();
